+++ sdk/resourcemanager/network/armnetwork/constants.go
@@ -11,7 +11,7 @@ package armnetwork
 
 const (
 	moduleName    = "armnetwork"
-	moduleVersion = "v2.2.1"
+	moduleVersion = "v2.3.0"
 )
 
 // Access - Access to be allowed or denied.